refactor(portal-frontend): hoist ASSERTION_MODE_OPTIONS + joinSeedUrl + drop dead roleBadge
Follow-up to SPEC-PORTAL-CONNECTOR-WIZARD-EXTRACT-001 (F-C2 partial, completes the hoist).

Three independent changes bundled because they all clean up `-kb-helpers.tsx`:

1. ASSERTION_MODE_OPTIONS (constant): moved to -connector-constants.ts.
   Wizard-only — no KB-tab consumer.

2. joinSeedUrl (function): moved to -connector-constants.ts.
   Wizard-only — used by both wizard pages + 1 test file. Kept in the
   constants file rather than splitting to its own -connector-helpers.ts
   because a single function does not warrant a separate file (the
   proliferation anti-pattern the rule itself warns against). The header
   comment documents that "constants AND a small helper here" is intentional.

3. roleBadge (function): removed entirely. Dead code — exported but
   no consumer found across the entire frontend src tree.

Test file renamed: __tests__/kb-helpers.test.ts → __tests__/connector-helpers.test.ts
(reflects the new home of the symbol it tests). Import path updated.

Removed `MultiSelectOption` import from -kb-helpers.tsx (was only used
by the moved ASSERTION_MODE_OPTIONS).

Net effect on cross-directory imports: BOTH wizard pages now consume
zero symbols from $kbSlug/-kb-helpers.tsx. They still consume
`CookieRow` and `ConnectorSummary` from $kbSlug/-kb-types.ts, which is
correct: those types are genuinely shared with KB-tab routes
(connectors.tsx) and components/knowledge/CookieRowsInput.tsx.

Verification:
- tsc --noEmit clean (only pre-existing tsconfig warning)
- eslint . clean (exit 0)
- vitest run __tests__/: 23/23 passing (joinSeedUrl test still green
  through the renamed file + new import path)
